git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6.7.1@4714 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/psd.c
@@ -2126,8 +2126,8 @@ static MagickBooleanType WritePSDImage(const ImageInfo *image_info,Image *image)
         (image->depth == 1) ? MagickTrue : MagickFalse;
       (void) WriteBlobMSBShort(image,(unsigned short)
         (monochrome != MagickFalse ? 1 : image->depth > 8 ? 16 : 8));
-      (void) WriteBlobMSBShort(image,monochrome != MagickFalse ? BitmapMode :
-        GrayscaleMode);
+      (void) WriteBlobMSBShort(image,(unsigned short)
+        (monochrome != MagickFalse ? BitmapMode : GrayscaleMode));
     }
   else
     {
